pyproject.toml
#[build-system]
#requires = ["setuptools"]
#build-backend = "setuptools.build_meta"

[build-system]
requires = ["setuptools>=42", "wheel", "nuitka", "toml"]
build-backend = "nuitka.distutils.Build"

[project]
name = "movx"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
    "clairmeta",
    "h2o_wave",
    "uvicorn",
    "click",
    "mkdocstrings-python",
    "pywebview",
]

[project.scripts]
movx = "movx.cli:main"

[nuitka]
nofollow-import-to = ["*.tests", "*.distutils"]